+++ lib/axios/signupActions.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-export const callSignup = (params, endpoint) => {
+export const getSignup = (params, endpoint) => {
   return axios.get(
     `/api/signup/v1/${endpoint}`,
     {
