+++ lib/axios/signupActions.test.js
@@ -1,9 +1,9 @@
-import * as actions from './signupAxiosActions';
+import * as actions from './signupActions';
 import axios from 'axios';
 jest.mock('axios');
 
 describe('actions/signUpAxiosActions', () => {
-  describe('callSignup', () => {
+  describe('getSignup', () => {
     const expected = {
       params: { user: 'username' },
       headers: {
@@ -16,7 +16,7 @@ describe('actions/signUpAxiosActions', () => {
     test('it should call axios get', () => {
       // mockAxios.post.mockImplementationOnce(() => Promise.resolve({}));
       const spy = jest.spyOn(axios, 'get');
-      actions.callSignup({ user: 'username' }, 'endpoint');
+      actions.getSignup({ user: 'username' }, 'endpoint');
       expect(spy).toHaveBeenCalled();
       expect(spy).toBeCalledWith('/api/signup/v1/endpoint', expected);
     });
